backend/user/serializers/user_serializer.py
from rest_framework import serializers
from user.models import CustomUser


class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = CustomUser
        fields = ['id', 'username', 'nickname', 'email', 'profile_image']
        read_only_fields = ['id']