backend/pyproject.toml
[tool.poetry]
name = "game-store"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.12"
pytest = "^8.3.3"
psycopg = "^3.1"
psycopg-binary = "^3.2.9"
factory-boy = "^3.3.1"
django = "^5.1.3"
djangorestframework = "^3.15.2"
django-extensions = "^3.2.3"
django-debug-toolbar = "^5.1.0"
pillow = "^11.2.1"	


[tool.poetry.group.dev.dependencies]
pytest-django = "^4.11.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
